% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{addCircleLayers}
\alias{addCircleLayers}
\title{Add Circle Layers}
\usage{
addCircleLayers(map, df_list, palette)
}
\arguments{
\item{map}{leaflet map object}

\item{df_list}{list of data.frames of well data split by reservoir name.}

\item{palette}{leaflet color mapping object which maps data values to colors according to a given palette.}
}
\value{
\code{addCircleLayers()} returns a leaflet proxy object
}
\description{
Creates a leaflet proxy with SL well and trajectory polygons.
}
