% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{analysisInfo}
\alias{analysisInfo}
\title{Current Session Information}
\usage{
analysisInfo(input, eval_date_time, sl_basename)
}
\arguments{
\item{input}{session input}

\item{eval_date_time}{date time object when report was generated}

\item{sl_basename}{name of original super location file}
}
\value{
\code{analysisInfo} returns a .csv with evaluation time, first_prod_date, reservoir, well_status,
hole_direction, normalized lateral length, units and superlocation file used to generate the report.
}
\description{
Obtain session info of reports created.
}
