% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{calcEPSG}
\alias{calcEPSG}
\title{Calculate the EPSG code needed to do calculations/buffers/transformations.
Will use the centeroid of the shapefile area given.}
\usage{
calcEPSG(shapefile)
}
\arguments{
\item{shapefile}{Simple Features object with CRS User Input: EPSG:4326}
}
\value{
WGS84 EPSG code representing the appropriate UTM with hemisphere
}
\description{
Calculate the EPSG code needed to do calculations/buffers/transformations.
Will use the centeroid of the shapefile area given.
}
