% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicesModule.R
\name{choicesServer}
\alias{choicesServer}
\title{Choices Server and Filtered Data}
\usage{
choicesServer(id, data, date)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{reactive named list containing the data needed for the map and filtered data.}

\item{date}{input from the date control for First Production Date used to filter.}
}
\value{
\code{choicesServer} returns a reactive filtered superlocation dataset based on selected UI choices.
}
\description{
Creates an internal reactive when sl dataset is received, which calculates the possible choices for
reservoir, well status and hole direction, then populates the corresponding UI components. The final
result is a reactive which is the subset of the sl data reflecting the selected choices.
}
