% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choicesModule.R
\name{choicesUI}
\alias{choicesUI}
\title{Choices UI}
\usage{
choicesUI(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used in the modules server function.}

\item{...}{Arguments supplied to selectInput controls.}
}
\value{
\code{choicesUI()} returns a list of three shiny selectizeInputs and an action button control.
}
\description{
Choices UI
}
