% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{computeSummary}
\alias{computeSummary}
\title{Compute Summary Data}
\usage{
computeSummary(
  filtered,
  multi,
  eur,
  nll = 10000,
  summary = FALSE,
  sumEUR = FALSE,
  magnitude = 1e+20
)
}
\arguments{
\item{filtered}{filtered subset of the SL dataset.}

\item{multi}{whether or not the shapefile has one or more polygons.}

\item{eur}{usfsrpmsc entity EUR data.}

\item{nll}{reactive Normalized Lateral Length}

\item{summary}{flag indicating whether or not data is used for the actual UI \code{DTOutput}}

\item{sumEUR}{flag indicating whether or not to generate data in wide or EUR formats.}

\item{magnitude}{at what point(magnitude) do represented numbers get changed to scientific notation.}
}
\value{
\code{computeSummary} returns one of three data sets depending on its flags. See above.
}
\description{
UI DT::datatable       summary =  TRUE & sumEUR = FALSE = tibble for DT EUR & EUR w/NLL
\code{is.null} = return empty tibble
}
\details{
Downloadable CSV(s)    summary = FALSE & sumEUR = FALSE = datatable in wide format
summary = FALSE & sumEUR = TRUE  = summary with EUR data
}
