% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{createDatatable}
\alias{createDatatable}
\title{Create datatable}
\usage{
createDatatable(summary_data, session)
}
\arguments{
\item{summary_data}{tibble created by \code{computeSummary()} function - summary data.}

\item{session}{current session info for namespacing}
}
\value{
\code{createDataTable()} returns \code{DT::datatable} object.
}
\description{
Wrapper for \code{DT::datatable} object. Adds a custom Shiny input button which allows the user filter options.
}
