% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{createNoDatatable}
\alias{createNoDatatable}
\title{Create NO datatable}
\usage{
createNoDatatable(session)
}
\arguments{
\item{session}{current session info for namespacing.}
}
\value{
\code{createNoDatatable()} returns an empty \code{DT::datatable} object, as a filler with custom message.
}
\description{
When summary_data is returned that has zero rows, need to retain the filter button.
}
