% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{ctrlGrouped}
\alias{ctrlGrouped}
\title{Leaflet Plugin - Control Grouped Layer}
\usage{
ctrlGrouped(source_path)
}
\arguments{
\item{source_path}{path where .css and .js scripts reside}
}
\value{
\code{ctrlGrouped} returns an HTML dependency the leaflet can implement for the plugin.
}
\description{
The leaflet map is currently limited to 2 types of group layers - Basegroups and Overlay.
The plugin allows new Layer Controls to be added to the map, such as variable map layers.
Load the JS and CSS files for groupedLayerControl library plugin, register the plugin function with
leaflet object.
}
\details{
https://stackoverflow.com/questions/67496042/leaflet-groupedlayercontrol-using-group-layers-in-r
}
