% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatableModule.R
\name{datatableServer}
\alias{datatableServer}
\title{Data table Server}
\usage{
datatableServer(id, filtered, eur, nll)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the modules UI functions.}

\item{filtered}{filtered subset of the SL dataset based on selected inputs.}

\item{eur}{usfsrpmsc EUR data.}

\item{nll}{reactive normalized lateral length input.}
}
\value{
\code{datatableServer} has side effects of creating and updating the summary data table, along with
custom filter input modal.
}
\description{
A \code{DT} datatable with reactive dependency on filtered SL dataset, also Normalized Lateral Length input.
Initial summary data is generated with user interaction, then any filters are applied and passed to the
\code{DT::renderDT} reactive output. Additional invalidation are made when any filters change or the
Normalized Lateral Length input changes, in which case the table utilizes a proxy which replaces the
current data set while retaining it's settings.
}
