% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatableModule.R
\name{datatableUI}
\alias{datatableUI}
\title{Data table User Interface}
\usage{
datatableUI(id, ...)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used in the modules server function.}

\item{...}{Additional arguments.}
}
\value{
\code{datatableUI} returns a list of:
\itemize{
\item \code{DT::DTOutput()} main datatable output.
\item \code{shiny::actionButton()} used to close the filter Modal.
\item \code{shiny::textOuput()} to display the active filter status.
}
}
\description{
Data table User Interface
}
