% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInputModule.R
\name{fileInputServer}
\alias{fileInputServer}
\title{File Upload}
\usage{
fileInputServer(id, wells, texas_counties, texas_grids)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{wells}{well data with simple feature geometries for well location and trajectory.}

\item{texas_counties}{\link{sf} with all Texas counties.}

\item{texas_grids}{\link{sf} with all Texas survey areas/grids.}
}
\value{
\code{fileInputServer()} returns a reactive list:
\itemize{
\item \code{shapefile} Original shapefile
\item \code{input$upload$name} Name of the shapefile
\item \code{center_lng/center_lat} Center points of the shapefile in longitude/latitude
\item \code{multi} Whether or not the shapefile has one or more polygons
\item \code{aoi_polys} Area of Interest Region \code{sf} polygons
\item \code{well_layers} Well Layers \code{data.frame}, where each well belongs to a specific region
\item \code{counties} County \code{sf} polygons
\item \code{grids} Grids \code{sf} polygons within the specific counties
\item \code{zoom} Zoom level for the leaflet based on total area
\item \code{got_data} Whether or not the input method contained data (file) or was manually selected
}
}
\description{
Processes an shapefile and calculates all wells within specified radius.
Check for existence of .prj projection file.
}
