% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileInputModule.R
\name{fileInputUI}
\alias{fileInputUI}
\title{File Input UI}
\usage{
fileInputUI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used in the modules server function.}
}
\value{
\code{fileInputUI()} returns a file upload control for input and text output element.
}
\description{
File Input UI
}
