% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getChoices}
\alias{getChoices}
\title{Create Input Choices}
\usage{
getChoices(header, ..., include_empty = FALSE)
}
\arguments{
\item{header}{calculated superlocation well data}

\item{...}{additional arguments, typically reservoir, status, drill and date inputs.}

\item{include_empty}{accounts for NA values}
}
\value{
list of possible combinations in the SL for:
\itemize{
\item \code{Reservoir}
\item \verb{Well Status}
\item \verb{Hole Direction}
\item \verb{First Production Date}
}
}
\description{
Used to calculate all combinations of "Reservoir", "Status" and "Drill" used to update interdependent
updateSelectize controls.
}
