% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletModule.R
\name{leafletServer}
\alias{leafletServer}
\title{Leaflet Server}
\usage{
leafletServer(id, data, filtered, eur, texas)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the modules UI functions.}

\item{data}{processed SL dataset.}

\item{filtered}{filtered subset of the SL dataset.}

\item{eur}{usfsrpmsc EUR data.}

\item{texas}{\link{sf} containing Texas state polygons.}
}
\value{
\code{leafletServer} returns a leaflet object or leaflet proxy object.
}
\description{
Two operations, initial leaflet generated is from the calculated superlocation data from file inputs.
Subsequent map is generated from the filtered choices dataset as a proxy update.
}
