% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafletModule.R
\name{leafletUI}
\alias{leafletUI}
\title{Leaflet UI}
\usage{
leafletUI(id, height = "50vh")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used in the modules server function.}

\item{height}{Height of the map in units view height vh.}
}
\value{
\code{leafletUI()} returns a leaflet UI element.
}
\description{
Leaflet UI
}
