% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{makeTSData}
\alias{makeTSData}
\title{Calculate Time Series Data}
\usage{
makeTSData(x)
}
\arguments{
\item{x}{entity or group level data.frame}
}
\value{
\code{makeTSData()} returns either entity or group level list with 3 data.frames:
\itemize{
\item \verb{time series(id, t, y}
TODO: created attrib
\item \code{metadata(id, group_id, source, type, product)}
\item \code{annotations}
}
}
\description{
Mutate rate time data to long format for \code{RscShiny::TSPlot}.
Any NA product values will calculate as 1.
}
