% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mapLabels}
\alias{mapLabels}
\title{Map Labels}
\usage{
mapLabels(wdf)
}
\arguments{
\item{wdf}{Filtered well data}
}
\value{
formatted HTML for well data popups on the leaflet map.
}
\description{
Map Labels
}
