% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{registerPlugin}
\alias{registerPlugin}
\title{Register Leaflet Plugin}
\usage{
registerPlugin(map, plugin)
}
\arguments{
\item{map}{leaflet map html widget}

\item{plugin}{plugin to add to map dependency}
}
\value{
leaflet map widget with added plugin.
}
\description{
Register Leaflet Plugin
}
