% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoirOverrideModule.R
\name{reservoirServer}
\alias{reservoirServer}
\title{Reservoir Override File Upload}
\usage{
reservoirServer(id, data)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{\code{reactive} sl_out object.}
}
\value{
\code{reservoirServer()} returns a reactive value containing the same data as the fileInputModule,
but with updated reservoir names. Also adds a copy of the original override CSV data.
}
\description{
Processes an Comma Separated Values file which contains name value pairs of Reservoir/Entity IDs. Used
to overwrite the original \code{Superlocation$well_layers} reservoir names to custom names.
}
