% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reservoirOverrideModule.R
\name{reservoirUI}
\alias{reservoirUI}
\title{Reservoir Override UI}
\usage{
reservoirUI(id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used in the modules server function.}
}
\value{
\code{reservoirUI()} returns an file upload control which accepts .csv input.
}
\description{
Reservoir Override UI
}
