% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{scaledVol}
\alias{scaledVol}
\title{Calculate Scaled EUR}
\usage{
scaledVol(vol, norm_lat_len, lat_len)
}
\arguments{
\item{vol}{Well volume.}

\item{norm_lat_len}{Hypothetical lateral length of the well.}

\item{lat_len}{Actual lateral length of the well.}
}
\value{
The scaled volume of the hypothesized lateral length compared to the original.
}
\description{
Normalize EUR|VOL by lateral length of well (EUR/ft), return NA if any values missing
eur, norm_lat_len = 10000 , ll = lower_perf - upper_perf
424/2567  # 16\% Horizontal well are missing upper or lower perfs
scaled is 'normalized' volume
}
