% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validCRS}
\alias{validCRS}
\title{Validate Coordinate Reference System}
\usage{
validCRS(shapefile)
}
\arguments{
\item{shapefile}{\code{sf} Simple Features object.}
}
\value{
logical operator indicating failure or success (TRUE).
}
\description{
When processing initial uploaded shapefile data, function checks to see if the shapefile contains a valid
coordinate reference system. If CRS is not valid, user is notified and file processing is halted.
}
