% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{validExtensions}
\alias{validExtensions}
\title{Validate files in superloc.zip}
\usage{
validExtensions(dir)
}
\arguments{
\item{dir}{temp dir generated by application where files are stored to be processed.}
}
\value{
logical operator indicating failure or success (TRUE).
}
\description{
When a user uploads a superlocation zipfile, it should contain the four file types needed:
\itemize{
\item .dbf  Feature Attributes
\item .prj  Projection Definition (CRS)
\item .shx  Indexed Geometry
\item .shp  Actual Shapefile Geometry
}
}
\details{
If any file types are missing: returns FALSE, user is notified which file extension is missing.
}
