% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{wellsInArea}
\alias{wellsInArea}
\title{Wells in Area}
\usage{
wellsInArea(wells, shapefile, aoi_region = 0)
}
\arguments{
\item{wells}{SF object, all wells          WGS84 EPSG:4326}

\item{shapefile}{single polygon SF object  WGS84 EPSG:4326}

\item{aoi_region}{radius in miles}
}
\value{
\code{wellsInArea()} returns the wells object with \code{aoi_region} column WGS84 EPSG:4326
}
\description{
Calculate the wells within a given radius area or polygon layer area.
Default is wells within the current polygon layer.
}
